Display hand update

+++ Blackjack/OutputFormatter.cs
@@ -11,7 +11,7 @@ namespace Blackjack
             var cards = hand.Cards;
             for (var i = 0; i < hand.Cards.Count; i++)
             {
-                stringBuilder.Append((int)cards[i].Rank); // TODO: if statment if face card return the value else int + 1 // check out Yatzy CategoryExtensions
+                stringBuilder.Append(cards[i].Rank); // TODO: cast as int; if statement if face card return the value else int + 1 // check out Yatzy CategoryExtensions
                 stringBuilder.Append(" of ");
                 stringBuilder.Append(cards[i].Suit);
                 stringBuilder.Append(", ");
